Hello.IntrovertServer (HelloWorld - Elixir v1.0.0)

The Introvert server enjoys quiet and prefers silence.

It will be polite about requesting silence, up to a point, but after that the introvert may get quite forceful about matters. The introvert has a finite temper and may get "propery radgy" (geordie english).

The introvert server always apologises for getting mad, and will clean up any mess that it has created. It doesn't hold a grudge but it also isn't very clever.

The introvert only speaks English.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_info(msg, state)

When the state is eight, introverts flip tables and kill things.

init(state)

Intialise function. State is a count of the amount of times introvert has told extrovert to be quiet. In a practical sense though it could be anything.

Parameters

state (integer): start at zero for maximum introvert tolerance

start_link(_)